removed legacy logger, websocket, constraint sample
