+++ nmmo/systems/item.py
@@ -7,7 +7,7 @@ from typing import Dict
 
 from nmmo.datastore.serialized import SerializedState
 from nmmo.lib.colors import Tier
-from nmmo.lib.log import EventCode
+from nmmo.lib.event_code import EventCode
 
 # pylint: disable=no-member
 ItemState = SerializedState.subclass("Item", [
@@ -182,19 +182,6 @@ class Equipment(Item):
 
     self.equipped.update(1)
     equip_slot.equip(self)
-
-    if self.config.LOG_MILESTONES and entity.is_player and self.config.LOG_VERBOSE:
-      for (label, level) in [
-        (f"{self.__class__.__name__}_Level", self.level.val),
-        ("Item_Level", entity.equipment.item_level),
-        ("Melee_Attack", entity.equipment.melee_attack),
-        ("Range_Attack", entity.equipment.range_attack),
-        ("Mage_Attack", entity.equipment.mage_attack),
-        ("Melee_Defense", entity.equipment.melee_defense),
-        ("Range_Defense", entity.equipment.range_defense),
-        ("Mage_Defense", entity.equipment.mage_defense)]:
-
-        self.realm.log_milestone(label, level, f'EQUIPMENT: {label} {level}')
 
   def _slot(self, entity):
     raise NotImplementedError
@@ -382,14 +369,7 @@ class Consumable(Item):
     assert self.listed_price == 0, "Listed item cannot be used"
     assert self._level(entity) >= self.level.val, "Entity's level is not sufficient to use the item"
 
-    self.realm.log_milestone(
-      f'Consumed_{self.__class__.__name__}', self.level.val,
-      f"PROF: Consumed {self.level.val} {self.__class__.__name__} "
-      f"by Entity level {entity.attack_level}",
-      tags={"player_id": entity.ent_id})
-
     self.realm.event_log.record(EventCode.CONSUME_ITEM, entity, item=self)
-
     self._apply_effects(entity)
     entity.inventory.remove(self)
     self.destroy()
