+++ tests/action/test_destroy_give_gold.py
@@ -10,7 +10,7 @@ from scripted import baselines
 
 RANDOM_SEED = 985
 
-LOGFILE = 'tests/action/test_destroy_give_gold.log'
+LOGFILE = None  # 'tests/action/test_destroy_give_gold.log'
 
 class TestDestroyGiveGold(ScriptedTestTemplate):
   # pylint: disable=protected-access,multiple-statements,no-member
@@ -28,8 +28,7 @@ class TestDestroyGiveGold(ScriptedTestTemplate):
     cls.ammo = { 1:Item.Whetstone, 2:Item.Arrow, 3:Item.Whetstone,
                  4:Item.Arrow, 5:Item.Whetstone, 6:Item.Arrow }
 
-    cls.config.LOG_VERBOSE = False
-    if cls.config.LOG_VERBOSE:
+    if LOGFILE:  # for debugging
       logging.basicConfig(filename=LOGFILE, level=logging.INFO)
 
   def test_destroy(self):
