+++ tests/render/test_load_replay.py
@@ -4,11 +4,11 @@ if __name__ == '__main__':
   import time
 
   # pylint: disable=import-error
-  from nmmo.render.render_client import WebsocketRenderer
+  from nmmo.render.render_client import DummyRenderer
   from nmmo.render.replay_helper import FileReplayHelper
 
   # open a client
-  renderer = WebsocketRenderer()
+  renderer = DummyRenderer()
   time.sleep(3)
 
   # load a replay: replace 'replay_dev.json' with your replay file
