+++ tests/testhelpers.py
@@ -111,14 +111,7 @@ class ScriptedAgentTestConfig(nmmo.config.Small, nmmo.config.AllGameSystems):
 
   __test__ = False
 
-  LOG_ENV = True
-
-  LOG_MILESTONES = True
-  LOG_EVENTS = False
-  LOG_VERBOSE = False
-
   PLAYER_DEATH_FOG = 5
-
   PLAYERS = [
     baselines.Fisher, baselines.Herbalist,
     baselines.Prospector,baselines.Carver, baselines.Alchemist,
